added backend
used express-generator to auto build an express server for the app so that an api can be implemented for songs. The react app was moved to the client folder where it contains its own development server. The express server will run at port 3001 so that it does not conflict with the react apps dev server at port 3000