changed stylings
making stylings to be set in css for consistency, trying to only apply styling in render if it depends on the state. Made the player look better, removed stylings from song card because a song card shouldnt have prior knowledge about how its going to be used it should just exists and the user should worry about how it is displayed where its implemented.